Extract sort test cases to SortTests
