+++ sorting/bogo_sort_test.rb
@@ -1,16 +1,11 @@
 require 'minitest/autorun'
+require_relative './sort_tests'
 require_relative './bogo_sort'
 
 class TestBogoSort < Minitest::Test
-  def test_sorted_array
-    input = [1, 2, 3, 4, 5]
-    expected = input.dup
-    assert_equal expected, input.bogosort
-  end
+  include SortTests
 
-  def test_reversed_array
-    input = [5, 4, 3, 2, 1]
-    expected = [1, 2, 3, 4, 5]
-    assert_equal expected, input.bogosort
+  def sort(input)
+    input.bogosort
   end
 end
